+++ app/src/js/components/Workflows/workflow.js
@@ -9,6 +9,7 @@ import config from '../../config';
 import { setWindowEditorRef } from '../../utils/browser';
 import Loading from '../LoadingIndicator/loading-indicator';
 import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
+import ErrorReport from '../Errors/report';
 
 class Workflows extends React.Component {
   constructor () {
@@ -26,20 +27,20 @@ class Workflows extends React.Component {
 
   renderReadOnlyJson (name, data) {
     return (
-        <Ace
-            mode='json'
-            theme={config.editorTheme}
-            name={`read-only-${name}`}
-            readOnly={true}
-            value={JSON.stringify(data, null, '\t')}
-            width='auto'
-            tabSize={config.tabSize}
-            showPrintMargin={false}
-            minLines={1}
-            maxLines={35}
-            wrapEnabled={true}
-            ref={setWindowEditorRef}
-        />
+      <Ace
+        mode='json'
+        theme={config.editorTheme}
+        name={`read-only-${name}`}
+        readOnly={true}
+        value={JSON.stringify(data, null, '\t')}
+        width='auto'
+        tabSize={config.tabSize}
+        showPrintMargin={false}
+        minLines={1}
+        maxLines={35}
+        wrapEnabled={true}
+        ref={setWindowEditorRef}
+      />
     );
   }
 
@@ -62,40 +63,40 @@ class Workflows extends React.Component {
     ];
 
     return (
-        <div className='page__component'>
-          <section className='page__section page__section__controls'>
-            <Breadcrumbs config={breadcrumbConfig} />
-          </section>
-          <h1 className='heading--large heading--shared-content with-description'>
-            {record.data ? record.data.long_name  : `...`}
-          </h1>
-          <section className='page__section'>
-            { record.inflight ? <Loading /> :
-                record.error ? <ErrorReport report={record.error} />
-                    : record.data ?
-                    <div>
-                      <div className='tab--wrapper'>
-                        <button className={'button--tab ' + (this.state.view === 'json' ? 'button--active' : '')}
-                                onClick={() => this.state.view !== 'json' && this.setState({ view: 'json' })}>JSON View</button>
-                      </div>
-                      <div>
-                        {this.state.view === 'list' ? this.renderList(record.data) : this.renderJson(record.data)}
-                      </div>
-                    </div> : null
-            }
-          </section>
-        </div>
+      <div className='page__component'>
+        <section className='page__section page__section__controls'>
+          <Breadcrumbs config={breadcrumbConfig} />
+        </section>
+        <h1 className='heading--large heading--shared-content with-description'>
+          {record.data ? record.data.long_name : '...'}
+        </h1>
+        <section className='page__section'>
+          { record.inflight ? <Loading />
+            : record.error ? <ErrorReport report={record.error} />
+              : record.data
+                ? <div>
+                  <div className='tab--wrapper'>
+                    <button className={'button--tab ' + (this.state.view === 'json' ? 'button--active' : '')}
+                      onClick={() => this.state.view !== 'json' && this.setState({ view: 'json' })}>JSON View</button>
+                  </div>
+                  <div>
+                    {this.state.view === 'list' ? this.renderList(record.data) : this.renderJson(record.data)}
+                  </div>
+                </div> : null
+          }
+        </section>
+      </div>
     );
   }
 
   renderJson (data) {
     return (
-        <ul>
-          <li>
-            <label>{data.name}</label>
-            {this.renderReadOnlyJson('recipe', data)}
-          </li>
-        </ul>
+      <ul>
+        <li>
+          <label>{data.name}</label>
+          {this.renderReadOnlyJson('recipe', data)}
+        </li>
+      </ul>
     );
   }
 }